move /etc/profile.d/00-pengwin.sh --> /etc/pengwin_conf.d/shellrc
plan is to have it sourced in bashrc and other shellrc files.
/etc/pengwin_conf.d/fishrc to be sourced in fishrc

this is part of a greater overall change to the shell environment
modifications that we make within the pengwin distribution (see
the similarly named branch in pengwin-setup Git repository). these
files will now be sourced on each execution of the shell binary,
as opposed to only on login as previously within /etc/profile.d

Signed-off-by: kim (grufwub) <grufwub@gmail.com>